split tokenizer tests into separate file; this is for illustration; not a CI thing
