+++ models/demos/deepseek_v3_d_p/tests/test_prefill_transformer.py
@@ -19,7 +19,6 @@ Parametrized over:
 
 import gc
 import json
-from pathlib import Path
 
 import pytest
 import torch
@@ -40,11 +39,9 @@ from models.demos.deepseek_v3_d_p.utils.transformer_helpers import (
     create_hf_model,
     download_infinitebench_subset,
     extract_tt_state_dict,
-    get_4d_causal_mask,
     load_and_compute_layer_by_layer,
     load_reference_cache,
     save_reference_cache,
-    tokenize_prompt_to_chat_template,
     tokenize_prompt_to_isl,
 )
 from tests.ttnn.utils_for_testing import comp_pcc
@@ -56,7 +53,7 @@ PCC_THRESHOLD = 0.99
 INFINITEBENCH_SUBSET_NAMES = {"passkey", "kv_retrieval", "longdialogue_qa_eng", "longbook_qa_eng"}
 
 
-@pytest.skipif(not is_blackhole(), reason="Requires Blackhole.")
+@pytest.mark.skipif(not is_blackhole(), reason="Requires Blackhole.")
 @pytest.mark.parametrize("return_kv_cache", [True], ids=["kv_cache"])
 @pytest.mark.parametrize("use_pretrained", [False, True], ids=["random", "pretrained"])
 @pytest.mark.parametrize(
@@ -489,131 +486,3 @@ def test_prefill_transformer(
     # Deferred PCC failure check (after timing report)
     if pcc_validation and has_pcc_failures:
         pytest.fail(f"PCC below {threshold} at: {pcc_failure_msg}")
-
-
-def test_tokenize_prompt_to_isl(tokenizer):
-    max_isl = 10
-    input_ids, attention_mask, tokens = tokenize_prompt_to_isl(
-        tokenizer, max_isl=max_isl, prompt_text="This is a test prompt.", debug=True
-    )
-
-    logger.debug(f"Input IDs: {input_ids}")
-    logger.debug(f"Attention Mask: {attention_mask}")
-    logger.debug(f"Tokens: {tokens}")
-
-    assert input_ids.shape == (1, max_isl), f"Expected input_ids shape (1, {max_isl}), got {input_ids.shape}"
-
-    torch.set_printoptions(threshold=float("inf"), edgeitems=3, precision=2, linewidth=200)
-    logger.debug(f"4D Causal Attention Mask shape:\n{get_4d_causal_mask(attention_mask, ignore_padding=True)}")
-    logger.debug(f"4D Causal Attention Mask Paddshape:\n{get_4d_causal_mask(attention_mask, ignore_padding=False)}")
-
-
-def test_tokenize_prompt_to_chat_template(tokenizer):
-    max_isl = 64
-    input_ids, tokens = tokenize_prompt_to_chat_template(
-        tokenizer,
-        max_isl=max_isl,
-        user_prompt="What is the capital of Serbia?",
-        system_prompt="You are a helpful assistant.",
-        debug=True,
-    )
-    logger.debug(f"Input IDs: {input_ids}")
-    logger.debug(f"Tokens: {tokens}")
-
-    assert input_ids.shape == (1, max_isl), f"Expected input_ids shape (1, {max_isl}), got {input_ids.shape}"
-
-
-@pytest.mark.parametrize(
-    "input_path",
-    [
-        Path("/tmp/r1/pretrained_abc_1k_isl1024_layers61_experts256.pt"),
-        Path(
-            "/DeepSeek-R1-0528-Cache/DeepSeek-R1-0528-Reference-prefill/pretrained_abc_1k_isl1024_layers61_experts256.pt"
-        ),
-        Path(
-            "/workspace/ds_output_all_sources_new_/norm_20260415_114721_mesh8x4_isl1024_L61_e256_cf32_gatehost_all_pretrained_abc_1k.pt"
-        ),
-    ],
-)
-def test_first_token_from_reference(input_path, model_path, config_only, tokenizer):
-    # Use weights_only=False since this is a trusted local file with custom objects
-    logger.info(f"{input_path=}")
-    data = torch.load(input_path, weights_only=False)
-
-    if "norm_output" in data:  # this is ttnn output; norm only
-        norm_output = data["norm_output"]
-        logger.info(f"{norm_output.shape=}")
-    elif "ref_snapshots" in data:  # this is torch output; all emb + layers + norm
-        logger.info(f"Number of reference snapshots: {len(data['ref_snapshots'])}")
-        assert len(data["ref_snapshots"]) == DeepSeekV3Config.NUM_LAYERS + 2  # token embedding + final RMS norm
-        norm_output = data["ref_snapshots"][-1]  # Last snapshot's tensor
-        logger.warning("Loaded data does not have 'norm_output' key, assuming last snapshot is final rms norm output")
-    else:
-        logger.warning(data)
-        raise ValueError("Loaded data format is unexpected and does not contain 'norm_output' or expected snapshots.")
-
-    #  Remove batch dimension if present
-    if norm_output.shape[0] == 1:
-        norm_output = norm_output.squeeze(0)
-    logger.success("Data loaded successfully.")
-
-    # LM HEAD loading
-    from models.demos.deepseek_v3.utils.config_helpers import sub_state_dict
-    from models.demos.deepseek_v3.utils.lazy_state_dict import LazyStateDict
-    from models.demos.deepseek_v3.utils.test_utils import dequantize_state_dict
-
-    lazy_sd = LazyStateDict(Path(model_path))
-    # lm_head_sd = sub_state_dict(lazy_sd, "model.embed_tokens.")
-    lm_head_sd = sub_state_dict(lazy_sd, "lm_head.")
-    lm_head_dequant = dequantize_state_dict(lm_head_sd, config_only)
-    lm_head_dequant_weight = lm_head_dequant.get("weight")
-    assert (
-        lm_head_dequant_weight.shape[0] == config_only.vocab_size
-    ), f"Expected lm_head_dequant_weight shape[0] to be {config_only.vocab_size}, got {lm_head_dequant_weight.shape[0]}"
-    logger.success("LM head weight loaded and dequantized successfully.")
-
-    # Apply LM HEAD
-    norm_output = norm_output.to(lm_head_dequant_weight.dtype)
-    # norm_output = ref_snapshots[-1][0,:,:].to(lm_head_dequant_weight.dtype)
-    logger.debug("Computing logits...")
-    with torch.no_grad():
-        # lm_head is just a linear layer: logits = hidden @ lm_head_weight.T
-        logits = torch.matmul(norm_output, lm_head_dequant_weight.T)  # Shape: [1, 1, vocab_size]
-    logger.debug(f"Logits shape: {logits.shape}")
-
-    # Apply sampling
-    # https://github.com/deepseek-ai/DeepSeek-V3/blob/main/inference/generate.py
-    def sample(logits, temperature: float = 1.0):
-        """
-        Samples a token from the logits using temperature scaling.
-
-        Args:
-            logits (torch.Tensor): The logits tensor for token predictions.
-            temperature (float, optional): Temperature for scaling logits. Defaults to 1.0.
-
-        Returns:
-            torch.Tensor: The sampled token.
-        """
-        logits = logits / max(temperature, 1e-5)
-        probs = torch.softmax(logits, dim=-1)
-        return probs.div_(torch.empty_like(probs).exponential_(1)).argmax(dim=-1)
-
-    index = -1  # Last token
-    for temperature in [0.0, 0.5, 1]:
-        first_token = sample(logits=logits[index].clone(), temperature=temperature)
-        token_text = tokenizer.decode([first_token.item()])
-        logger.info(f"{index} First token (temperature={temperature}): {first_token.item()} -> {repr(token_text)}")
-
-    last_logit = logits[-1, :].clone()
-    top_k = 5
-    top_logits, top_indices = torch.topk(last_logit, top_k, dim=-1)
-    probs = torch.softmax(last_logit, dim=-1)
-
-    for i, (token_id, logit_value) in enumerate(zip(top_indices.tolist(), top_logits.tolist())):
-        token_text = tokenizer.decode([token_id])
-        prob = probs[token_id].item()
-
-        logger.info(
-            f"{i+1}. Token ID: {token_id:6d} | Logit: {logit_value:8.4f} | "
-            f"Prob: {prob:8.6f} | Text: {repr(token_text)}"
-        )
